Use one python file for each test category
